+++ src/app/pages/collection/foreword/collection-foreword.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 
 import { TextChangerComponent } from '@components/text-changer/text-changer.component';
-import { TrustHtmlPipe } from '@pipes/trust-html-pipe';
+import { TrustHtmlPipe } from '@pipes/trust-html.pipe';
 import { CollectionForewordPage } from './collection-foreword.page';
 import { CollectionForewordPageRoutingModule } from './collection-foreword-routing.module';