+++ src/app/pages/collection/introduction/collection-introduction.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 
 import { TextChangerComponent } from '@components/text-changer/text-changer.component';
-import { TrustHtmlPipe } from '@pipes/trust-html-pipe';
+import { TrustHtmlPipe } from '@pipes/trust-html.pipe';
 import { CollectionIntroductionPage } from './collection-introduction.page';
 import { CollectionIntroductionPageRoutingModule } from './collection-introduction-routing.module';