+++ src/app/pages/content/content.module.ts
@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
 import { ContentGridComponent } from '@components/content-grid/content-grid.component';
-import { TrustHtmlPipe } from '@pipes/trust-html-pipe';
+import { TrustHtmlPipe } from '@pipes/trust-html.pipe';
 import { ContentPage } from './content.page';
 import { ContentPageRoutingModule } from './content-routing.module';