+++ src/app/pages/elastic-search/elastic-search.module.ts
@@ -6,7 +6,7 @@ import { IonicModule } from '@ionic/angular';
 import { DateHistogramComponent } from '@components/date-histogram/date-histogram.component';
 import { ElasticHitPagePathPipe } from '@pipes/elastic-hit-page-path.pipe';
 import { ElasticHitQueryparamsPipe } from '@pipes/elastic-hit-queryparams.pipe';
-import { TrustHtmlPipe } from '@pipes/trust-html-pipe';
+import { TrustHtmlPipe } from '@pipes/trust-html.pipe';
 import { ElasticSearchPageRoutingModule } from './elastic-search-routing.module';
 import { ElasticSearchPage } from './elastic-search.page';