+++ src/app/pages/media-collection/media-collection.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
-import { TrustHtmlPipe } from '@pipes/trust-html-pipe';
+import { TrustHtmlPipe } from '@pipes/trust-html.pipe';
 import { MediaCollectionPageRoutingModule } from './media-collection-routing.module';
 import { MediaCollectionPage } from './media-collection.page';